+++ src/fixtures/functions/extractData.ts
